api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.31
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
